@@ -7,106 +7,113 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-native-community/datetimepicker": "2.2.2",
     "@react-native-community/geolocation": "^2.0.2",
-    "axios": "^0.19.0",
-    "firebase": "^6.3.1",
+    "@react-native-community/masked-view": "^0.1.7",
+    "@react-native-community/netinfo": "^5.5.1",
+    "@react-navigation/drawer": "^5.1.1",
+    "@react-navigation/native": "^5.0.9",
+    "@react-navigation/stack": "^5.1.1",
+    "axios": "^0.19.2",
+    "firebase": "^7.9.3",
     "immutable": "^4.0.0-rc.12",
-    "lodash": "^4.17.11",
+    "lodash": "^4.17.15",
     "logdecor": "^1.0.2",
-    "lottie-ios": "3.0.3",
-    "lottie-react-native": "3.1.1",
+    "lottie-ios": "3.1.6",
+    "lottie-react-native": "3.3.2",
     "moment": "^2.24.0",
-    "native-base": "2.13.1",
-    "papaparse": "^4.6.3",
+    "native-base": "2.13.8",
+    "papaparse": "^5.1.1",
     "prop-types": "^15.7.2",
-    "react": "16.6.3",
-    "react-native": "^0.59.10",
-    "react-native-android-location-enabler": "^1.1.0",
+    "react": "16.13.0",
+    "react-native": "^0.61.5",
+    "react-native-android-location-enabler": "^1.2.0",
     "react-native-android-wifi": "0.0.41",
-    "react-native-animatable": "^1.3.2",
+    "react-native-animatable": "^1.3.3",
     "react-native-audio-record": "^0.2.2",
-    "react-native-audio-recorder-player": "^2.3.0",
+    "react-native-audio-recorder-player": "^2.5.0",
     "react-native-background-timer": "2.1.1",
     "react-native-check-box": "^2.1.7",
     "react-native-collapsible": "^1.5.1",
     "react-native-communications": "^2.2.1",
     "react-native-contacts-wrapper": "^0.2.4",
     "react-native-countdown-circle": "^3.0.1",
-    "react-native-country-picker-modal": "^0.8.0",
+    "react-native-country-picker-modal": "^1.10.0",
     "react-native-custom-qr-codes": "^2.0.0",
     "react-native-datepicker": "^1.7.2",
     "react-native-datepicker-dialog": "0.0.9",
-    "react-native-device-info": "2.2.2",
+    "react-native-device-info": "5.5.3",
     "react-native-document-picker": "^3.2.4",
-    "react-native-draggable-flatlist": "^1.1.7",
+    "react-native-draggable-flatlist": "^2.3.0",
     "react-native-easy-toast": "1.2.0",
-    "react-native-elements": "^1.0.0",
+    "react-native-elements": "^1.2.7",
     "react-native-elevated-view": "^0.0.6",
     "react-native-fabric": "^0.5.2",
-    "react-native-firebase": "^5.5.6",
-    "react-native-flash-message": "^0.1.13",
+    "react-native-firebase": "^5.6.0",
+    "react-native-flash-message": "^0.1.15",
     "react-native-flexi-radio-button": "^0.2.2",
-    "react-native-fs": "^2.13.3",
-    "react-native-geolocation-service": "^3.0.0",
-    "react-native-gesture-handler": "^1.5.2",
+    "react-native-fs": "^2.16.6",
+    "react-native-geolocation-service": "^4.0.0",
+    "react-native-gesture-handler": "^1.6.0",
     "react-native-html-to-pdf": "^0.8.0",
     "react-native-image-picker": "0.28.0",
     "react-native-image-placeholder": "^1.0.14",
-    "react-native-image-to-pdf": "^1.1.0",
+    "react-native-image-to-pdf": "^1.2.0",
     "react-native-keyboard-aware-scroll-view": "latest",
     "react-native-location-satellites": "^1.0.5",
-    "react-native-maps": "^0.24.2",
+    "react-native-maps": "^0.26.1",
     "react-native-marquee": "^0.3.2",
     "react-native-material-dropdown": "^0.11.1",
-    "react-native-material-textfield": "^0.12.0",
-    "react-native-modal": "^11.1.0",
-    "react-native-modal-datetime-picker": "^6.0.0",
-    "react-native-permissions": "^2.0.3",
-    "react-native-phone-input": "^0.2.1",
-    "react-native-qrcode-svg": "^5.2.0",
-    "react-native-reanimated": "^1.4.0",
-    "react-native-responsive-screen": "^1.2.0",
-    "react-native-screens": "^1.0.0-alpha.23",
-    "react-native-select-multiple": "^2.0.0",
-    "react-native-share": "^1.1.3",
+    "react-native-material-textfield": "^0.16.1",
+    "react-native-modal": "^11.5.4",
+    "react-native-modal-datetime-picker": "^8.5.1",
+    "react-native-permissions": "^2.0.9",
+    "react-native-phone-input": "^0.2.4",
+    "react-native-qrcode-svg": "^6.0.3",
+    "react-native-reanimated": "^1.7.0",
+    "react-native-responsive-screen": "^1.4.0",
+    "react-native-safe-area-context": "^0.7.3",
+    "react-native-screens": "^2.2.0",
+    "react-native-select-multiple": "^2.1.0",
+    "react-native-share": "^3.0.0",
     "react-native-side-drawer": "^1.2.5",
-    "react-native-simple-radio-button": "^2.7.3",
+    "react-native-simple-radio-button": "^2.7.4",
     "react-native-sliding-calendar-view": "^0.1.3",
     "react-native-sound": "^0.11.0",
-    "react-native-spinkit": "^1.3.0",
-    "react-native-svg": "^9.9.9",
-    "react-native-system-setting": "^1.7.2",
+    "react-native-spinkit": "^1.5.0",
+    "react-native-svg": "^11.0.1",
+    "react-native-system-setting": "^1.7.4",
     "react-native-table-component": "^1.2.1",
-    "react-native-timeago": "^0.4.0",
+    "react-native-timeago": "^0.5.0",
     "react-native-timer": "^1.3.6",
     "react-native-vector-icons": "^6.6.0",
-    "react-native-version-number": "^0.3.5",
+    "react-native-version-number": "^0.3.6",
     "react-native-view-pdf": "^0.10.1",
-    "react-native-view-shot": "^2.6.0",
+    "react-native-view-shot": "^3.1.2",
     "react-native-zoom-image": "^0.1.2",
-    "react-navigation": "^4.0.10",
-    "react-navigation-drawer": "^2.3.3",
-    "react-navigation-redux-helpers": "^3.0.2",
-    "react-navigation-stack": "^1.10.3",
-    "react-navigation-tabs": "^2.6.2",
-    "react-redux": "^6.0.0",
-    "redux": "^4.0.1",
+    "react-navigation": "^4.2.2",
+    "react-navigation-drawer": "^2.4.2",
+    "react-navigation-redux-helpers": "^4.0.1",
+    "react-navigation-stack": "^2.2.2",
+    "react-navigation-tabs": "^2.8.2",
+    "react-redux": "^7.2.0",
+    "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
-    "redux-persist": "^5.10.0",
-    "redux-persist-filesystem-storage": "^1.3.2",
+    "redux-persist": "^6.0.0",
+    "redux-persist-filesystem-storage": "^2.1.0",
     "redux-thunk": "^2.3.0",
-    "rn-fetch-blob": "0.10.15",
+    "rn-fetch-blob": "0.12.0",
     "rn-progress-loader": "^1.0.8",
     "toggle-switch-react-native": "^2.1.0"
   },
   "devDependencies": {
-    "babel-jest": "^25.0.0",
-    "babel-plugin-module-resolver": "^3.2.0",
+    "babel-jest": "^25.1.0",
+    "babel-plugin-module-resolver": "^4.0.0",
     "babel-plugin-transform-remove-console": "^6.9.4",
-    "jest": "^25.0.0",
-    "jetifier": "^1.6.4",
-    "metro-react-native-babel-preset": "0.51.1",
-    "react-test-renderer": "16.6.3"
+    "jest": "^25.1.0",
+    "jetifier": "^1.6.5",
+    "metro-react-native-babel-preset": "0.58.0",
+    "react-test-renderer": "16.13.0"
   },
   "jest": {
     "preset": "react-native"
