@@ -7,7 +7,7 @@
     4. staging --- For Staging server //For staging server Media upload url same as uat
 */
 
-let urlType = 'apidev';
+let urlType = 'apiuat';
 
 const oyeURL = `${urlType}.oyespace.com`; //Validation/
 //Image Domains
